Added segments impl and test

+++ src/main/java/com/configcat/SegmentCondition.java
@@ -5,16 +5,18 @@ import com.google.gson.annotations.SerializedName;
 public class SegmentCondition {
 
     @SerializedName(value = "s")
-    private Double segmentIndex;
+    private int segmentIndex;
 
     @SerializedName(value = "c")
     private int segmentComparator;
 
-    public Double getSegmentIndex() {
+    public int getSegmentIndex() {
         return segmentIndex;
     }
 
     public int getSegmentComparator() {
         return segmentComparator;
     }
+
+
 }
